chore: Split `jsonpath`
