+++ dozer-sql/src/lib.rs
@@ -1,11 +1,4 @@
-extern crate core;
-
 // Re-export sqlparser
 pub use sqlparser;
 
-pub mod jsonpath;
 pub mod pipeline;
-
-#[macro_use]
-extern crate pest_derive;
-extern crate pest;
